[Mrunal]Refactor: Multiperson dupication single line consol command check in person add
